Cocher la case 32 bit lrclock dans les blocs ip i2s. Corriger le facteur de scaling du rate des blocs (parametre 0x20 ===> 8 dans vitis). Petite detection des oiseaux vite fait.

+++ ZYBO/ZYBO_PROJET_S4/VITIS_WORKSPACE/ZYBO_PROJET_S4/src/FIFO_FFT_driver.h
@@ -39,8 +39,8 @@ int XLFifoInit(XLlFifo *InstancePtr, u16 DeviceId);
 #define FIFO_CONFIG_ID		XPAR_AXI_FIFO_0_DEVICE_ID	  //THe FIFO only configures the FFT block
 
 
-extern u32 ReceiveTempBuffer[MAX_DATA_BUFFER_SIZE]; //buffer for the interrupt data storage
+extern short ReceiveTempBuffer[MAX_DATA_BUFFER_SIZE]; //buffer for the interrupt data storage
 extern char flagDonnes;
-extern u32 BufferMain[MAX_DATA_BUFFER_SIZE];
+extern int BufferMain[MAX_DATA_BUFFER_SIZE];
 
 #endif /* FIFO_H_ */
